Extract and document a type.
